+++ CHANGELOG.md
@@ -12,6 +12,8 @@ project adheres to
 
 * Added axum example (PR #118, thanks @vbrandl).
 * Updated rsass to 0.27.0 and base64 to 0.21.0.
+* Updated dependencies in examples: actix-web 4.2.1, axum 0.6.2,
+  env_logger 0.10.0,
 
 
 ## Releaase 0.15.0